# Conflicts:
#	dev/Screen/User/MailBox.js

+++ snappymail/v/0.0.0/app/libraries/RainLoop/Providers/Storage/FileStorage.php
@@ -2,12 +2,14 @@
 
 namespace RainLoop\Providers\Storage;
 
+use RainLoop\Providers\Storage\Enumerations\StorageType;
+
 class FileStorage implements \RainLoop\Providers\Storage\IStorage
 {
 	/**
 	 * @var string
 	 */
-	private $sDataPath;
+	protected $sDataPath;
 
 	/**
 	 * @var bool
@@ -27,26 +29,25 @@ class FileStorage implements \RainLoop\Providers\Storage\IStorage
 	}
 
 	/**
-	 * @param \RainLoop\Model\Account|string|null $oAccount
+	 * @param \RainLoop\Model\Account|string|null $mAccount
 	 */
-	public function Put($oAccount, int $iStorageType, string $sKey, string $sValue) : bool
+	public function Put($mAccount, int $iStorageType, string $sKey, string $sValue) : bool
 	{
-		return false !== \file_put_contents(
-			$this->generateFileName($oAccount, $iStorageType, $sKey, true), $sValue);
+		$sFileName = $this->generateFileName($mAccount, $iStorageType, $sKey, true);
+		return $sFileName && false !== \file_put_contents($sFileName, $sValue);
 	}
 
 	/**
-	 * @param \RainLoop\Model\Account|string|null $oAccount
+	 * @param \RainLoop\Model\Account|string|null $mAccount
 	 * @param mixed $mDefault = false
 	 *
 	 * @return mixed
 	 */
-	public function Get($oAccount, int $iStorageType, string $sKey, $mDefault = false)
+	public function Get($mAccount, int $iStorageType, string $sKey, $mDefault = false)
 	{
 		$mValue = false;
-		$sFileName = $this->generateFileName($oAccount, $iStorageType, $sKey);
-		if (\file_exists($sFileName))
-		{
+		$sFileName = $this->generateFileName($mAccount, $iStorageType, $sKey);
+		if ($sFileName && \file_exists($sFileName)) {
 			$mValue = \file_get_contents($sFileName);
 		}
 
@@ -54,14 +55,13 @@ class FileStorage implements \RainLoop\Providers\Storage\IStorage
 	}
 
 	/**
-	 * @param \RainLoop\Model\Account|string|null $oAccount
+	 * @param \RainLoop\Model\Account|string|null $mAccount
 	 */
-	public function Clear($oAccount, int $iStorageType, string $sKey) : bool
+	public function Clear($mAccount, int $iStorageType, string $sKey) : bool
 	{
 		$mResult = true;
-		$sFileName = $this->generateFileName($oAccount, $iStorageType, $sKey);
-		if (\file_exists($sFileName))
-		{
+		$sFileName = $this->generateFileName($mAccount, $iStorageType, $sKey);
+		if ($sFileName && \file_exists($sFileName)) {
 			$mResult = \unlink($sFileName);
 		}
 
@@ -69,26 +69,14 @@ class FileStorage implements \RainLoop\Providers\Storage\IStorage
 	}
 
 	/**
-	 * @param \RainLoop\Model\Account|string $oAccount
+	 * @param \RainLoop\Model\Account|string $mAccount
 	 */
-	public function DeleteStorage($oAccount) : bool
+	public function DeleteStorage($mAccount) : bool
 	{
-		$sPath = $this->generateFileName($oAccount,
-			\RainLoop\Providers\Storage\Enumerations\StorageType::USER, 'xxx', false, true);
-
-		if (!empty($sPath) && \is_dir($sPath))
-		{
+		$sPath = $this->generateFileName($mAccount, StorageType::CONFIG, '', false, true);
+		if ($sPath && \is_dir($sPath)) {
 			\MailSo\Base\Utils::RecRmDir($sPath);
 		}
-
-		$sPath = $this->generateFileName($oAccount,
-			\RainLoop\Providers\Storage\Enumerations\StorageType::CONFIG, 'xxx', false, true);
-
-		if (!empty($sPath) && \is_dir($sPath))
-		{
-			\MailSo\Base\Utils::RecRmDir($sPath);
-		}
-
 		return true;
 	}
 
@@ -100,66 +88,59 @@ class FileStorage implements \RainLoop\Providers\Storage\IStorage
 	/**
 	 * @param \RainLoop\Model\Account|string|null $mAccount
 	 */
-	public function generateFileName($mAccount, int $iStorageType, string $sKey, bool $bMkDir = false, bool $bForDeleteAction = false) : string
+	protected function generateFileName($mAccount, int $iStorageType, string $sKey, bool $bMkDir = false, bool $bForDeleteAction = false) : string
 	{
-		if (null === $mAccount)
-		{
-			$iStorageType = \RainLoop\Providers\Storage\Enumerations\StorageType::NOBODY;
-		}
-
 		$sEmail = $sSubEmail = '';
-		if ($mAccount instanceof \RainLoop\Model\Account)
-		{
-			$sEmail = $mAccount->ParentEmailHelper();
-			if ($this->bLocal && $mAccount->IsAdditionalAccount() && !$bForDeleteAction)
+		if (null === $mAccount) {
+			$iStorageType = StorageType::NOBODY;
+		} else if ($mAccount instanceof \RainLoop\Model\Account) {
+			$sEmail = $mAccount instanceof \RainLoop\Model\AdditionalAccount ? $mAccount->ParentEmail() : $mAccount->Email();
+			if ($this->bLocal && $mAccount instanceof \RainLoop\Model\AdditionalAccount && !$bForDeleteAction)
 			{
 				$sSubEmail = $mAccount->Email();
 			}
-		}
-
-		if (\is_string($mAccount) && empty($sEmail))
-		{
+		} else if (\is_string($mAccount) && empty($sEmail)) {
 			$sEmail = $mAccount;
 		}
 
-		$sEmail = \preg_replace('/[^a-z0-9\-\.@]+/i', '_', $sEmail);
-		$sSubEmail = \preg_replace('/[^a-z0-9\-\.@]+/i', '_', $sSubEmail);
-
-		$sTypePath = $sKeyPath = '';
+		$sFilePath = '';
 		switch ($iStorageType)
 		{
+			case StorageType::NOBODY:
+				$sFilePath = $this->sDataPath.'/__nobody__/'.\sha1($sKey ?: \time());
+				break;
+			case StorageType::SIGN_ME:
+				$sSubEmail = '.sign_me';
+			case StorageType::CONFIG:
+				if (empty($sEmail)) {
+					return '';
+				}
+				if (\is_dir("{$this->sDataPath}/cfg")) {
+					\SnappyMail\Upgrade::FileStorage($this->sDataPath);
+				}
+				$aEmail = \explode('@', $sEmail ?: 'nobody@unknown.tld');
+				$sDomain = \trim(1 < \count($aEmail) ? \array_pop($aEmail) : '');
+				$sFilePath = $this->sDataPath
+					.'/'.\RainLoop\Utils::fixName($sDomain ?: 'unknown.tld')
+					.'/'.\RainLoop\Utils::fixName(\implode('@', $aEmail) ?: '.unknown')
+					.'/'.($sSubEmail ? \RainLoop\Utils::fixName($sSubEmail).'/' : '')
+					.($sKey ? \RainLoop\Utils::fixName($sKey) : '');
+				break;
 			default:
-			case \RainLoop\Providers\Storage\Enumerations\StorageType::USER:
-			case \RainLoop\Providers\Storage\Enumerations\StorageType::NOBODY:
-				$sTypePath = 'data';
-				$sKeyPath = \md5($sKey);
-				$sKeyPath = \substr($sKeyPath, 0, 2).'/'.$sKeyPath;
-				break;
-			case \RainLoop\Providers\Storage\Enumerations\StorageType::CONFIG:
-				$sTypePath = 'cfg';
-				$sKeyPath = \preg_replace('/[_]+/', '_', \preg_replace('/[^a-zA-Z0-9\/]/', '_', $sKey));
-				break;
+				throw new \Exception("Invalid storage type {$iStorageType}");
 		}
 
-		$sFilePath = '';
-		if (\RainLoop\Providers\Storage\Enumerations\StorageType::NOBODY === $iStorageType)
-		{
-			$sFilePath = $this->sDataPath.'/'.$sTypePath.'/__nobody__/'.$sKeyPath;
-		}
-		else if (!empty($sEmail))
-		{
-			$sFilePath = $this->sDataPath.'/'.$sTypePath.'/'.
-				\str_pad(\rtrim(\substr($sEmail, 0, 2), '@'), 2, '_').'/'.$sEmail.'/'.
-				(0 < \strlen($sSubEmail) ? $sSubEmail.'/' : '').
-				($bForDeleteAction ? '' : $sKeyPath);
-		}
-
-		if ($bMkDir && !$bForDeleteAction && !empty($sFilePath) && !\is_dir(\dirname($sFilePath)))
+		if ($bMkDir && !empty($sFilePath) && !\is_dir(\dirname($sFilePath)))
 		{
 			if (!\mkdir(\dirname($sFilePath), 0700, true))
 			{
 				throw new \RainLoop\Exceptions\Exception('Can\'t make storage directory "'.$sFilePath.'"');
 			}
+		}
+
+		// CleanupSignMeData
+		if (StorageType::SIGN_ME === $iStorageType && $sKey && 0 === \random_int(0, 25) && \is_dir($sFilePath)) {
+			\MailSo\Base\Utils::RecTimeDirRemove(\is_dir($sFilePath), 3600 * 24 * 30); // 30 days
 		}
 
 		return $sFilePath;
